Move interfaces to a new folder
