+++ src/Microsoft.Diagnostics.Runtime.Tests/src/ByReferenceTests.cs
@@ -5,6 +5,7 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using Microsoft.Diagnostics.Runtime.Interfaces;
 using Xunit;
 
 namespace Microsoft.Diagnostics.Runtime.Tests
