+++ src/Microsoft.Diagnostics.Runtime/ClrAppDomain.cs
@@ -3,6 +3,7 @@
 // See the LICENSE file in the project root for more information.
 
 using Microsoft.Diagnostics.Runtime.Implementation;
+using Microsoft.Diagnostics.Runtime.Interfaces;
 using System.Collections.Generic;
 using System.Collections.Immutable;
