+++ src/Microsoft.Diagnostics.Runtime/ClrHeap.cs
@@ -3,6 +3,7 @@
 // See the LICENSE file in the project root for more information.
 
 using Microsoft.Diagnostics.Runtime.Implementation;
+using Microsoft.Diagnostics.Runtime.Interfaces;
 using System;
 using System.Buffers;
 using System.Collections.Generic;
