+++ src/Microsoft.Diagnostics.Runtime/ClrInstanceField.cs
@@ -4,6 +4,7 @@
 
 using Microsoft.Diagnostics.Runtime.DacInterface;
 using Microsoft.Diagnostics.Runtime.Implementation;
+using Microsoft.Diagnostics.Runtime.Interfaces;
 using System;
 using System.Diagnostics;
 using System.Reflection;
