+++ src/Microsoft.Diagnostics.Runtime/ClrStackInteriorRoot.cs
@@ -2,6 +2,8 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
+using Microsoft.Diagnostics.Runtime.Interfaces;
+
 namespace Microsoft.Diagnostics.Runtime
 {
     public sealed class ClrStackInteriorRoot : IClrStackRoot
