+++ src/Microsoft.Diagnostics.Runtime/ClrThread.cs
@@ -4,6 +4,7 @@
 
 using Microsoft.Diagnostics.Runtime.DacInterface;
 using Microsoft.Diagnostics.Runtime.Implementation;
+using Microsoft.Diagnostics.Runtime.Interfaces;
 using System.Collections.Generic;
 
 namespace Microsoft.Diagnostics.Runtime
