+++ src/Microsoft.Diagnostics.Runtime/ComCallableWrapper.cs
@@ -3,6 +3,7 @@
 // See the LICENSE file in the project root for more information.
 
 using Microsoft.Diagnostics.Runtime.DacInterface;
+using Microsoft.Diagnostics.Runtime.Interfaces;
 using System.Collections.Immutable;
 
 namespace Microsoft.Diagnostics.Runtime
