+++ src/Microsoft.Diagnostics.Runtime/Interfaces/IClrException.cs
@@ -4,7 +4,7 @@
 
 using System.Collections.Immutable;
 
-namespace Microsoft.Diagnostics.Runtime
+namespace Microsoft.Diagnostics.Runtime.Interfaces
 {
     public interface IClrException
     {
