+++ src/Microsoft.Diagnostics.Runtime/Interfaces/IClrHeap.cs
@@ -5,7 +5,7 @@
 using System.Collections.Generic;
 using System.Collections.Immutable;
 
-namespace Microsoft.Diagnostics.Runtime
+namespace Microsoft.Diagnostics.Runtime.Interfaces
 {
     public interface IClrHeap
     {
