+++ src/Microsoft.Diagnostics.Runtime/Interfaces/IClrJitManager.cs
@@ -1,6 +1,6 @@
 ﻿using System.Collections.Generic;
 
-namespace Microsoft.Diagnostics.Runtime
+namespace Microsoft.Diagnostics.Runtime.Interfaces
 {
     public interface IClrJitManager
     {
