+++ src/Microsoft.Diagnostics.Runtime/Interfaces/IClrMethod.cs
@@ -5,7 +5,7 @@
 using System.Collections.Immutable;
 using System.Reflection;
 
-namespace Microsoft.Diagnostics.Runtime
+namespace Microsoft.Diagnostics.Runtime.Interfaces
 {
     public interface IClrMethod
     {
