+++ src/Microsoft.Diagnostics.Runtime/Interfaces/IClrRuntime.cs
@@ -6,7 +6,7 @@ using System;
 using System.Collections.Generic;
 using System.Collections.Immutable;
 
-namespace Microsoft.Diagnostics.Runtime
+namespace Microsoft.Diagnostics.Runtime.Interfaces
 {
     public interface IClrRuntime : IDisposable
     {
