+++ src/Microsoft.Diagnostics.Runtime/Interfaces/IClrSegment.cs
@@ -4,7 +4,7 @@
 
 using System.Collections.Generic;
 
-namespace Microsoft.Diagnostics.Runtime
+namespace Microsoft.Diagnostics.Runtime.Interfaces
 {
     public interface IClrSegment
     {
