+++ src/Microsoft.Diagnostics.Runtime/Interfaces/IClrStackRoot.cs
@@ -2,7 +2,7 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-namespace Microsoft.Diagnostics.Runtime
+namespace Microsoft.Diagnostics.Runtime.Interfaces
 {
     public interface IClrStackRoot : IClrRoot
     {
