+++ src/Microsoft.Diagnostics.Runtime/Interfaces/IClrType.cs
@@ -7,7 +7,7 @@ using System.Collections.Generic;
 using System.Collections.Immutable;
 using System.Reflection;
 
-namespace Microsoft.Diagnostics.Runtime
+namespace Microsoft.Diagnostics.Runtime.Interfaces
 {
     public interface IClrType : IEquatable<IClrType>
     {
